requirements.txt
flask>=3.0.1,<4.0.0
flask-login>=0.6.3,<1.0.0
flask-sqlalchemy>=3.1.1,<4.0.0
werkzeug>=3.0.1,<4.0.0
scipy>=1,<2.0.0
numpy>=2.2,<3.0.0
matplotlib>=3.10,<4.0.0
pillow>=10.2.0,<11.0.0
SQLAlchemy>=2.0.25,<3.0.0
WTForms>=3.1.1,<4.0.0
email_validator>=2.1.0,<3.0.0
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0
flower>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
